requirements.txt
absl-py==1.3.0
alembic==1.7.7
astunparse==1.6.3
cachetools==4.2.4
charset-normalizer==2.1.1
click==8
colorama==0.4.5
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.48.2
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.31
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
